Applied "matchPDict() boost feature" patch
With this patch, finding all the matches of a 3.3M 32-mers dictionary in the
full Human genome (+ and - strands of all chromosomes) is about 2.5x faster
than before (will take between 20 minutes and 2 hours depending on the machine
and the number of matches found). This puts matchPDict() at the same level as
the Vmatch software (http://www.vmatch.de/) for a dictionary of this size.
Memory footprint for matchPDict() is still reasonable: between 500M and 800M
for the Aho-Corasick tree built from the 3.3M 32-mers dictionary. Building
this tree is still very fast (2 or 3 minutes) (Vmatch needs 60G of disk space
to build all its suffix arrays, don't know how long it takes for this, don't
know what's the memory footprint either when they are loaded into memory but
it looks like it is several Gigas).
There is still room for making matchPDict() even faster (10%, maybe 20%).
Reminder: matchPDict() still only works with a dictionary of DNA patterns
where all the patterns have the same number of nucleotides (Vmatch doesn't
have this kind of limitations). This can be addressed in the future depending
on the kind of applications we want to target (our main goal for now is to be
able to work with Solexa data).

Version bump.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Biostrings@29486 bc3139a8-67e5-0310-9ffc-ced21a209358